Add an abstract function that can change they column data type
Implementation of #197
Thanks @endinc!
